fix: divide components
