--- src/components/header/Header.js
@@ -1,5 +1,0 @@
-function Header() {
-  return <div></div>;
-}
-
-export default Header;
